styl: :art: Move fetchApps to the top of the Bar
